+++ apiserver/paasng/paas_wl/resources/actions/archive.py
@@ -19,11 +19,9 @@ to the current version of the project delivered to anyone in the future.
 """Archive related deploy functions"""
 import logging
 
-from paas_wl.cnative.specs.procs import get_proc_specs
 from paas_wl.monitoring.app_monitor.managers import make_bk_monitor_controller
-from paas_wl.platform.applications.models import Release
-from paas_wl.workloads.processes.controllers import AppProcessesController, CNativeProcController
-from paasng.platform.applications.constants import ApplicationType
+from paas_wl.workloads.processes.controllers import get_proc_ctl
+from paas_wl.workloads.processes.shim import ProcessManager
 from paasng.platform.applications.models import ModuleEnvironment
 
 logger = logging.getLogger(__name__)
@@ -41,20 +39,8 @@ class ArchiveOperationController:
 
     def stop_all_processes(self):
         """Stop all processes"""
-        if self.env.application.type == ApplicationType.CLOUD_NATIVE:
-            self._stop_cnative_all_processes()
-        else:
-            self._stop_all_processes()
-
-    def _stop_all_processes(self):
-        """普通应用，插件应用等根据 DB 中的配置，逐个停止进程"""
-        ctl = AppProcessesController(env=self.env)
-        release: Release = Release.objects.get_latest(self.env.wl_app)
-        for proc_type in release.get_procfile().keys():
+        ctl = get_proc_ctl(env=self.env)
+        lister = ProcessManager(env=self.env)
+        for proc_spec in lister.list_processes_specs():
+            proc_type = proc_spec["name"]
             ctl.stop(proc_type=proc_type)
-
-    def _stop_cnative_all_processes(self):
-        """云原生应用取线上现有进程，逐个停止"""
-        ctl = CNativeProcController(env=self.env)
-        for spec in get_proc_specs(self.env):
-            ctl.stop(proc_type=spec.name)
