+++ apiserver/paasng/tests/paas_wl/utils/wl_app.py
@@ -25,7 +25,7 @@ from django.utils.crypto import get_random_string
 
 from paas_wl.platform.applications.models import Build, Release, WlApp
 from paas_wl.platform.applications.models.config import Config
-from paas_wl.workloads.processes.models import Instance
+from paas_wl.workloads.processes.entities import Instance
 from tests.utils.auth import create_user
